Updated openstack/openstack Project: openstack/neutron  b1c2cfdc8b977743c4ac1b8bc0023168e1067d70
Update pylint disable list to pass pylint 1.7.1 checks

Recenly pylint version was bumped to 1.7.1 in global-requirements.txt.
Pylint 1.7.1 adds more checks and breaks the neutron pep8 job.

This commit adds new checks to the disabled list to avoid the gate
breakage. Individual disabled tests can be fixed later.

The new disabled list is compatible with pylint 1.4.5,
so we can merge this regardless of pylint bump is reverted or not.

Closes-Bug: #1696403
Change-Id: If6228d0626413049b82f9d75bcdf3bea7ddacde5