Updated openstack/openstack Project: openstack-infra/tripleo-ci  b25a048716a963d22c9d263eed6ba6d2a5756220
Add masquerading for controllers so that they can reach DNS

In OVB overcloud node use undercloud as default route. In OOOQ
environment a masqerading rule is set in [1], but in non-OOOQ
jobs this is not done and overcloud nodes can't reach DNS servers.

This patch adds a iptables masquerading rule similarly as OOOQ does.

Related-Bug: #1696094

[1] http://git.openstack.org/cgit/openstack/tripleo-quickstart-extras/tree/roles/baremetal-prep-overcloud/templates/allow-traffic-for-controller.sh.j2

Change-Id: I3961958e9d9c548827ba83301170f1e9c9d2203d