Updated openstack/openstack Project: openstack-infra/puppet-jenkins  3a1e5525fd5cb15ba34cfbeb15e45dd2113d2a73
Remove unnecessary beaker tests

We don't need beaker to check that every single file and package is
installed. It is safe to assume that if puppet ran successfully then it
was also able to successfully find and install packages and apply config
files. Rather than update the tests to reflect new names of Java
packages on xenial, we can reduce maintenance costs by just removing the
tests.

It's important to keep tests that check that the service is running and
working correctly, because it is possible for puppet to claim success
but for the application to not be working.

Change-Id: Ifd41b858812638191049f415bbdf78c1f3145e46