Updated openstack/openstack Project: openstack-infra/puppet-jenkins  e3f67e7941c0d4bd44ff9e7fb38e4a397ffcd78f
Use the default jenkins password for JJB

This module does not create a jenkins user within jenkins, so there is
no way its password could have been set to 'secret'. Trying to use this
as JJB's auth credentials was causing the JJB exec to fail. This patch
grabs the default admin password out of /var and uses that for JJB.

Change-Id: Ifa60398acbc494dcfecd0cabaa8c829590c1744e