Updated openstack/openstack Project: openstack-infra/puppet-logstash  567ae88a960ec822197d68bfa1a68e58bf53506f
Fix curator cron command string

The puppet syntax check complains about the '\%' in the cron command
string, but the % still does need to be escaped for the cron to work
properly. We double escape it here in the puppet string which will
result in a properly single-escaped command string in the crontab.

Change-Id: I3d97463a611c43ded12ddb788e9da5e13466e0c5