Updated openstack/openstack Project: openstack-infra/puppet-logstash  38c66fc220ccbed4c4fe7e29ec7e19cb286a5583
Remove unnecessary beaker tests

We do not need to check every package and file with beaker. It is safe
to assume that if puppet ran successfully then it also was able to find
and install every package and was able to apply every config file.

We still keep the service checks because it is possible for puppet to
declare a service to be running successfully only for it to die later.

Change-Id: I313fe7907b165e4c064cc37cf0c75051aabf103e