Updated openstack/openstack Project: openstack/python-troveclient  a5b7be4bed27bf5b9b2466b79cb73ac95665a8a1
Add user-list to OSC

This change adds database support to the python-openstackclient
project for the user-list command.

The trove command user-list is now:
    openstack database user list

Note: Only looking up an instance by ID works. I think it has
something to do with not using osc-lib.find_resource because we
use our own find_resource in the client manager's list call.

Change-Id: If8a4b97b91c661e400dfcac0fa4922ed359ce32a
Partially-Implements: trove-support-in-python-openstackclient