Updated openstack/openstack Project: openstack/python-troveclient  2aefbc59c74aa9e726837cc87705077d8eea2724
Add database-list to OSC

This change adds database support to the python-openstackclient
project for the database-list command.

The trove command database-list is now:
    openstack database list

To confirm the databases list command I had to create an instance
and specify a database name like how it shows in the user-guide[1]
on step 2.

[1] http://docs.openstack.org/user-guide/create-db.html

Change-Id: I4f4682eff1addc3fc837f5468939ddd96572e238
Partially-Implements: trove-support-in-python-openstackclient