Updated openstack/openstack Project: openstack-infra/puppet-lodgeit  f807e4de7d62320bec6ed5cd0ca8a987fa9bc843
Remove unnecessary beaker tests

It is good enough to assume that if puppet ran successfully then it also
was able to sucessfully find and install packages and apply config
files. For the sake of maintainability, this patch removes beaker tests
for individual files and packages. It also removes a check for a cron
that was already deleted from the manifest.

Change-Id: Ia8b29684213bc8c96d23fb9dd53baf8e65c85362