Updated openstack/openstack Project: openstack/tempest  a194cf1d91cb21616e1b4dec3c37d05c46486fe1
Replace assertions with more specific ones

This patch adds the already implemented assert(Not)Empty functions
instead of generic assert functions where assert(Not)Equal or
assertGreater(Equal) is used to check the state of a container.
This change makes the code and the error messages more readable,
therefore it improves maintainability a bit.

Change-Id: Ib7df910c7853141bcf2effd89f8fba864fc6bbb5