Updated openstack/openstack Project: openstack/nova  e845d66a1baeacaf7cbb3f5cd3fc3e75961ec372
Adopts keystoneauth with glance client.

Migrate Glance client to Keystonesuth to make it
consistent with Cinder and Neutron clients.

Note:  We still use the api_servers variable, rather than
fetching endpoints from the service catalog.
This change is needed to enable the use of service token
with Glance client.

Related to blueprint use-service-tokens-pike
Change-Id: I02be31bbe2de54c69210934b234ed380daca4fda
Co-Authored-By: Pushkar Umaranikar <pushkar.umaranikar@intel.com>