Updated openstack/openstack Project: openstack/manila  403aafacb624e882803b9b951b205d307d4481b3
Change example value in docs for CephFS snapshots

The documentation for the CephFS Native Manila driver would
show the configuration stanza with the
"cephfs_snapshot_support" option set to "true". The snapshot
support in CephFS is experimental and comes with a list of
caveats, so this patch changes the value shown in the example
to be "false" as well as adds a warning after the command to
enable snapshots in the doc explaining the risks involved with
CephFS snapshots.

Change-Id: I4f2c7ebf213c85abd25da078e3dd4999c5e7333d