Updated openstack/openstack Project: openstack/nova  447f2432902ca411ddc8f07ecb42719aa4993885
Fix the race condition with novnc

The first RFP negotiation message potentially arrived
earlier, in which case the message will be swallowed
by the recv() which receive response for the http
request. This commit is to ensure only the response
is removed from the socket buffer, so it won't impact
the following RFP negotiation.

Change-Id: I100e140acbc2d981f7b98e12b3e9ae02844f41fd
Closes-Bug: #1695844