Updated openstack/openstack Project: openstack/cinder  623a22fc4961874443abdef94840aaa6c5498853
Fix response code for update and delete API in volume attachments

The normal response code for update or delete API described in volume
attachments api-ref is 202, but the actual response code from the curl
request is 200. This patch is to fix the response codes of these two
APIs in api-ref.

Change-Id: I6eb43f63bb504cb64b12f55de759779058b2979d