Updated openstack/openstack Project: openstack/nova  bbbb043b992c465f75ebcfbc1d2c653936d1a6e7
Remove old service version check for mitaka

This removes a service version check, that was added in Newton, where
we needed to be graceful about build requests that may not have been
created by a Newton API node. This has long since left our support
envelope and can be removed (instead of fixed for multiple cells).

Change-Id: I75bdf973106b2ef3861b5c8b6ae742c719769fc4