Updated openstack/openstack Project: openstack/heat  0e67fb4c6eb8dab106e68e3ca8aa96d4d12532f6
Don't always validate property values in ResourceChain

Generally with nested stacks, when validating the parent stack we don't
validate the property *values* in the nested stack as we recurse down (i.e.
the strict_validate flag is set to False on the nested stack). We validate
the values when we come to actually create the nested stack.

ResourceChain is an exception - the line of code to do this was commented
out without explanation during initial development of the resource. This
has two consequences:

* Resources in the child stack always have their property values validated
  during the validation of the root stack, as well as when the child stack
  is created. (Property values should be examined only in the latter
  validation pass).
* Since ea2673fb9a04588e0e294d159945d075657b9112, any grandchild stacks
  (i.e. where a member of the ResourceChain is itself a StackResource of
  type) will not be validated at all during the validation of the root
  stack. (They should be still validated, though property values will not
  be examined.)

This change makes ResourceChain behave like other StackResources.

Change-Id: Icc9c698a9f787b211af5dc2e190f03eb8d841610
Related-Bug: #1675589