Updated openstack/openstack Project: openstack/nova  96d059ff38aaab1038cb18a4ced9810de6c2bfc0
remove ugly local import

To avoid cyclic dependency a local import is introduced in
I1399d15285677aab4dd79704d21fbaf2bb5e1361 but it can be
removed with a small refactoring as proposed here.

The original get_nw_info_for_instance() is kept in nova.compute.utils
as a proxy to avoid changes in a lot of other source files. That
removal will be done in a subsequent patch.

Change-Id: I32231126483d419358acc40d353edd45af744a1f