Updated openstack/openstack Project: openstack/heat  9e410ec4dd2d3ea41ec31a70f2c42e2cc619a12f
Get the deployment group resources again after CREATE_COMPLETE

We seem to get the list of group resources for signaling (there
is possibility that resource_id is not set for some resources)
and then use the same list to get the deployments. It would be
good to get the resources again after they are created.

Change-Id: I908d1d13abe8e59a65308e883591abca2b1c7a9a
Partial-Bug: #1625921