Updated openstack/openstack Project: openstack/python-novaclient  c45bb4dfc78a4163ea276e64ebeb51fdbcfcc495
Error out if nic auto or none are not alone

Previously, the client would allow a boot command containing, for
example, --nic auto,tag=foo, giving the impression that the auto nic
would receive a tag. This is not the case. Instead of silently
ignoring the tag argument in such a case, this patch makes the client
error out with a warning message.

Change-Id: Id7ef67435fa7a3f49c9abcb751f5fa9458388f13
Closes-bug: 1695088