Updated openstack/openstack Project: openstack/neutron  172046863854f2d7241da3185c79d6a1f80a1285
DVR: Fix DVR Router snat ports and gateway ports host binding issue

DVR snat ports and gateway ports are not bound to any host
and so we don't see the snat namespace getting created.

The issue is the _build_routers_list in l3_dvr_db.py is not called due
to the inheritance order.

Change-Id: I56f9de31524aeef262cf2a78be3abf8487c21a12
Closes-Bug: #1695101