Updated openstack/openstack Project: openstack-infra/shade  ab8a3d51fc4cca27137796bf69dcef2265eab1ce
RESTify create_server

In order to create passing tags to the nics parameter of create_server,
we need to support microversions. In order to support microversions, we
need to be using REST. Yolanda needs to pass tags to the nics parameter,
so let's go ahead and transition create_server since we want to anyway.

Tests are modified to remove an extra GET call that novaclient makes
that we don't actually need.

A follow up to this will add microversion discovery, and then will add
support for 'tags' being in the nics list.

Change-Id: I04c3b08d80cf496dabec8ce7000ca5fb679fe3b9