Updated openstack/openstack Project: openstack/horizon  2be0a22863750f8fd02ebf5aeb855d5bff5cf54d
Use same form for Create Router in network topology

Previously the create router form in network topology panel
used its own form template and by some mistake no message was
shown in the right side. It seems the reason of the original
template is to use a dedicate form submit url, but it can be
done by overriding the form class.

This commit overrides the create router form and deletes
the form templates specific to network-topology panel.
This reduces the maintenance cost of templates.

In addition, a page now remains in the network topology panel
when a router creation fails.

Change-Id: I78aeead952d5bcd31606cb3d561ccfa8f095ec36