Updated openstack/openstack Project: openstack/neutron  2d752108d04767054d4f54a09eec35f1c612add9
Fix test_dvr_gateway_host_binding_is_set

Router scheduling was happening before the schedule call
so sometimes the router was being scheduled to a different
host than the test was expecting.

This fixes it by explicitly setting the router to not be HA
and calls schedule before adding interfaces that would trigger
scheduling.

Closes-Bug: #1698058
Change-Id: I6c6319f1da72546a9d21af198daf9a2a67e6dec8