Updated openstack/openstack Project: openstack/tripleo-heat-templates  cea62f129cadbee58480765b3f9b6b0894ca418f
Modify PreNetworkConfig config inline with role-specific parameters

Existing host_config_and_reboot.role.j2.yaml is done in ocata to
configure kernel args. This can be enhanced with use of role-specific
parameters, which is done in the current patch. The earlier method is
deprecated and will be removed in Q releae.
Implements: blueprint ovs-2-6-dpdk

Change-Id: Ib864f065527167a49a0f60812d7ad4ad12c836d1